change button & ecamples to typescript
