--- app/frontend/src/components/Common/SearchField/SimpleSearchButton.js
@@ -1,19 +1,0 @@
-import { LitElement, html } from 'lit';
-
-import Styles from '../../../../stylesheets/object/component/simple-search-button.scss';
-
-export default class SearchButton extends LitElement {
-  static get styles() {
-    return [Styles];
-  }
-
-  constructor() {
-    super();
-  }
-
-  render() {
-    return html`<button class="btn"></button>`;
-  }
-}
-
-customElements.define('search-button', SearchButton);
